requirements.txt
Flask==1.0.2
gunicorn==19.9.0
beautifulsoup4==4.6.3
line-bot-sdk==1.8.0
imgurpython==1.1.7
urllib3==1.25.9
